% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizar_analisis_discriminante.R
\name{analisis_discriminante_csv}
\alias{analisis_discriminante_csv}
\title{Realizar análisis discriminante lineal desde un archivo CSV}
\usage{
analisis_discriminante_csv(data, clase, variables)
}
\arguments{
\item{data}{Un data frame que contiene los datos del archivo CSV.}

\item{clase}{El nombre de la columna que se utilizará como variable de clase.}

\item{variables}{Un vector de nombres de columna que se utilizarán como variables predictoras.}
}
\value{
Un objeto de modelo LDA.
}
\description{
Esta función realiza un análisis discriminante lineal (LDA) utilizando los datos proporcionados en un archivo CSV.
}
