% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_medidas_precision.R
\name{calcular_medidas_precision}
\alias{calcular_medidas_precision}
\title{Calcular medidas de precisión}
\usage{
calcular_medidas_precision(conf_mat)
}
\arguments{
\item{conf_mat}{Una matriz de confusión.}
}
\value{
Un data frame con las medidas de precisión.
}
\description{
Esta función calcula medidas de precisión como la exactitud, sensibilidad, especificidad, precisión y recall a partir de una matriz de confusión.
}
