% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_diagrama_disperion.R
\name{crear_diagrama_disperion}
\alias{crear_diagrama_disperion}
\title{Crear un diagrama de dispersión basado en análisis discriminante lineal}
\usage{
crear_diagrama_disperion(data, variables)
}
\arguments{
\item{data}{Un data frame que contiene los datos de entrada.}

\item{variables}{Un vector de nombres de columna que se utilizarán como variables predictoras en el modelo LDA.}
}
\description{
Esta función crea un diagrama de dispersión utilizando las funciones discriminantes como ejes.
}
