% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_analisis_discriminante.R
\name{generar_analisis_discriminante}
\alias{generar_analisis_discriminante}
\title{Función principal para el análisis discriminante y la generación de gráficos}
\usage{
generar_analisis_discriminante(archivo_csv, variables)
}
\arguments{
\item{archivo_csv}{La ruta del archivo CSV que contiene los datos de entrada.}

\item{variables}{Un vector de nombres de columna que se utilizarán como variables predictoras en el análisis.}
}
\value{
Una lista con los objetos de modelo, predicciones, modelo de validación cruzada y medidas de precisión.
}
\description{
Esta función principal carga los paquetes necesarios, lee los datos del archivo CSV, realiza el análisis discriminante, la validación cruzada y genera gráficos.
}
