% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_matriz_confusion.R
\name{obtener_matriz_confusion}
\alias{obtener_matriz_confusion}
\title{Obtener la matriz de confusi贸n}
\usage{
obtener_matriz_confusion(prediccion, clase)
}
\arguments{
\item{prediccion}{Un vector de predicciones obtenido a partir del modelo.}

\item{clase}{Un vector que contiene las clases reales de los datos.}
}
\value{
Una matriz de confusi贸n.
}
\description{
Esta funci贸n crea la matriz de confusi贸n a partir de las predicciones y las clases reales.
}
