% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizar_validacion_cruzada.R
\name{realizar_validacion_cruzada}
\alias{realizar_validacion_cruzada}
\title{Realizar validación cruzada de análisis discriminante lineal}
\usage{
realizar_validacion_cruzada(data, variables)
}
\arguments{
\item{data}{Un data frame que contiene los datos de entrada.}

\item{variables}{Un vector de nombres de columna que se utilizarán como variables predictoras.}
}
\value{
Un objeto de modelo de validación cruzada.
}
\description{
Esta función realiza validación cruzada en el conjunto de datos utilizando el método de análisis discriminante lineal (LDA).
}
